chore: remake code
