+++ keadm/cmd/keadm/app/cmd/reset_others.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 /*
 Copyright 2022 The KubeEdge Authors.
